updated many files

+++ utils/cloudinary.video.js
@@ -1,5 +1,5 @@
-import { v2 as cloudinary } from 'cloudinary';
-import fs from 'fs';
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 import dotenv from "dotenv";
 
 dotenv.config();
@@ -9,32 +9,34 @@ cloudinary.config({
   api_key: process.env.CLOUDINARY_API_KEY,
   api_secret: process.env.CLOUDINARY_API_SECRET,
 });
+// utils/cloudinary.js
+const uploadVideoOnCloudinary = async (localFilePath) => {
+  if (!localFilePath) return null;
 
+  try {
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "video",
+      folder: "videos",
+        chunk_size: 6000000, // VERY IMPORTANT
+      eager_async: true,   // async processing
+    });
 
+    console.log(response)
 
-const uploadVideoOnCloudinary= async  (localFilePath)=>{
-    try {
-        if (!localFilePath)return null;
-        
-        const response =cloudinary.uploader.upload(localFilePath,{
-            resource_type: "video",
-            
-        });
-        fs.unlinkSync(localFilePath)
-        return response
-
-
-    } catch (error) {
-      console.error("Cloudinary upload failed:", error);
-
-        try {
-                  fs.unlinkSync(localFilePath);
-              } catch (unlinkError) {
-                  console.error("Failed to delete local file:", unlinkError);
-              }
-               return null;
-
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
     }
-} 
+
+    return response;
+  } catch (error) {
+    console.error("Cloudinary upload failed:", error);
+
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+
+    return null;
+  }
+};
 
 export {uploadVideoOnCloudinary}
